Update 9/4 3:01 am
Almost the final cut. I finished cleaning out the redundant and useless code.
It now uses the datastore to store all lesson data. It still uses the Lesson class to pull the data from the txt files then pulls the it from the object and inputs it into a datastore entity.

Also added the DGoods class. This defines the object that holds all the data being passed to the templates at render. This cleans up the calls to the render function and makes the code easier to navigate.